+++ scratch.js
@@ -3,14 +3,16 @@ const readFile = require('./tests/lib/_cachedPage');
 // const wtf = require('./builds/wtf_wikipedia');
 // const wtf = require('./build');
 let options = {
-  // sections: false,
+  sections: false,
   // paragraphs: false,
   // sentences: true,
-  // lists: false,
   // title: false,
+  lists: false,
   // links: false,
-  // images: false,
-  // tables: false,
+  templates: false,
+  citations: true
+// images: false,
+// tables: false,
 };
 // (async () => {
 //   const doc = await wtf.fetch(`United Kingdom`, 'en');
@@ -18,13 +20,8 @@ let options = {
 // })();
 
 
-// let doc = readFile('United-Kingdom');
+let doc = readFile('United-Kingdom');
 // console.log(doc.sentences(0).html(options));
 
-var doc = readFile('royal_cinema');
-var data = doc.json({
-  templates: false,
-  images: false,
-  infoboxes: true
-});
-console.log(data);
+// var doc = readFile('royal_cinema');
+console.log(doc.html(options));