+++ src/document/toHtml.js
@@ -1,3 +1,11 @@
+const setDefaults = require('../lib/setDefaults');
+const defaults = {
+  title: true,
+  infoboxes: true,
+  headers: true,
+  sections: true,
+  links: true,
+};
 // we should try to make this look like the wikipedia does, i guess.
 const softRedirect = function(doc) {
   let link = doc.redirectTo();
@@ -13,6 +21,7 @@ const softRedirect = function(doc) {
 
 //turn a Doc object into a HTML string
 const toHtml = function(doc, options) {
+  options = setDefaults(options, defaults);
   let data = doc.data;
   let html = '';
   html += '<!DOCTYPE html>\n';
@@ -20,7 +29,7 @@ const toHtml = function(doc, options) {
   html += '<head>\n';
   //add page title
   if (options.title === true && data.title) {
-    html += '<title>' + data.title + ' - Wikipedia' + '</title>\n';
+    html += '<title>' + data.title + '</title>\n';
   }
   html += '</head>\n';
   html += '<body>\n';
@@ -30,18 +39,40 @@ const toHtml = function(doc, options) {
     html += softRedirect(doc);
     return html + '\n</body>\n</html>'; //end it here.
   }
-
-  //add header
-  if (options.title === true && data.title) {
-    html += '  <h1>' + data.title + '</h1>\n';
-  }
   //render infoboxes (up at the top)
   if (options.infoboxes === true) {
     html += doc.infoboxes().map(i => i.html(options)).join('\n');
   }
   //render each section
-  if (options.sections === true || options.paragraphs === true || options.sentences === true) {
+  if (options.sections === true && (options.paragraphs === true || options.sentences === true)) {
     html += data.sections.map(s => s.html(options)).join('\n');
+  }
+  //default off
+  if (options.citations === true) {
+    html += '<h2>References</h2>';
+    html += doc.citations().map((o, i) => {
+      if (o.data && o.data.url && o.data.title) {
+        let str = o.data.title;
+        if (options.links === true) {
+          str = `<a href="${o.data.url}">${str}</a>`;
+        }
+        return `<div class="citation">${i} - ${str} </div>`;
+      } else if (o.data.encyclopedia) {
+        return `<div class="citation">${i} - ${o.data.encyclopedia}</div>`;
+      } else if (o.data.title) { //cite book, etc
+        let str = o.data.title;
+        if (o.data.author) {
+          str += o.data.author;
+        }
+        if (o.data.first && o.data.last) {
+          str += o.data.first + ' ' + o.data.last;
+        }
+        return `<div class="citation">${i} - ${str}</div>`;
+      } else if (o.inline) {
+        return `<div class="citation">${i} - ${o.inline.html()}</div>`;
+      }
+      return '';
+    }).join('\n');
   }
   html += '</body>\n';
   html += '</html>';