+++ src/document/toJson.js
@@ -1,19 +1,17 @@
 const setDefaults = require('../lib/setDefaults');
 const redirects = require('./redirects');
-
 const defaults = {
-  title: true,
-  pageID: true,
-  categories: true,
-  coordinates: true,
+  infoboxes: true,
+  headers: true,
   sections: true,
+  pageID: true,
 };
 
 //an opinionated output of the most-wanted data
 const toJSON = function(doc, options) {
   options = setDefaults(options, defaults);
   let data = {};
-  if (options.title) {
+  if (options.headers || options.title) {
     data.title = doc.options.title || doc.title();
   }
   if (options.pageID && doc.options.pageID) {
@@ -21,9 +19,6 @@ const toJSON = function(doc, options) {
   }
   if (options.categories) {
     data.categories = doc.categories();
-  }
-  if (options.citations) {
-    data.citations = doc.citations();
   }
   if (options.coordinates) {
     data.coordinates = doc.coordinates();
@@ -39,21 +34,24 @@ const toJSON = function(doc, options) {
   if (options.sections) {
     data.sections = doc.sections().map(i => i.json(options));
   }
+  if (doc.isRedirect() === true) {
+    data.isRedirect = true;
+    data.redirectTo = redirects.parse(doc.wiki);
+    data.sections = [];
+  }
 
   //these are default-off
   if (options.plaintext) {
     data.plaintext = doc.plaintext(options);
+  }
+  if (options.citations) {
+    data.citations = doc.citations();
   }
   if (options.markdown) {
     data.markdown = doc.markdown(options);
   }
   if (options.html) {
     data.html = doc.html(options);
-  }
-  if (doc.isRedirect() === true) {
-    data.isRedirect = true;
-    data.redirectTo = redirects.parse(doc.wiki);
-    data.sections = [];
   }
   return data;
 };