+++ src/section/toJson.js
@@ -1,46 +1,43 @@
 const setDefaults = require('../lib/setDefaults');
 const defaults = {
-  title: true,
+  headers: true,
   depth: true,
   paragraphs: true,
-  links: true,
-  text: true,
-  formatting: true,
-  dates: true,
-  tables: true,
-  lists: true,
-  templates: true,
   images: true,
+  tables: true,
+  templates: true,
+  lists: true,
 };
 //
 const toJSON = function(section, options) {
   options = setDefaults(options, defaults);
   let data = {};
-  if (options.title) {
+  if (options.headers === true) {
     data.title = section.title();
   }
-  if (options.depth) {
+  if (options.depth === true) {
     data.depth = section.depth;
   }
   //these return objects
   if (options.paragraphs === true) {
     data.paragraphs = section.paragraphs().map(p => p.json(options));
   }
-  if (options.sentences === true) {
-    data.sentences = section.sentences().map(s => s.json(options));
-  }
-  if (options.images) {
+  if (options.image === true) {
     data.images = section.images().map(img => img.json(options));
   }
   //more stuff
-  if (options.tables) {
+  if (options.tables === true) {
     data.tables = section.tables().map(t => t.json(options));
   }
-  if (options.templates) {
+  if (options.templates === true) {
     data.templates = section.templates();
   }
-  if (options.lists) {
+  if (options.lists === true) {
     data.lists = section.lists().map(list => list.json(options));
+  }
+  //default off
+  if (options.sentences === true) {
+    data.sentences = section.sentences().map(s => s.json(options));
   }
   return data;
 };